+++ packages/plugins/expiration/CHANGELOG.md
@@ -1,5 +1,12 @@
 # learn-card-core
 
+## 1.1.61
+
+### Patch Changes
+
+-   Updated dependencies [[`cfabf6686a0233ed89de6201a70c01598c5ab298`](https://github.com/learningeconomy/LearnCard/commit/cfabf6686a0233ed89de6201a70c01598c5ab298)]:
+    -   @learncard/vc-plugin@1.2.0
+
 ## 1.1.60
 
 ### Patch Changes