+++ services/learn-card-network/learn-cloud-service/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @learncard/network-brain-service
 
+## 2.3.19
+
+### Patch Changes
+
+-   Updated dependencies [[`cfabf6686a0233ed89de6201a70c01598c5ab298`](https://github.com/learningeconomy/LearnCard/commit/cfabf6686a0233ed89de6201a70c01598c5ab298)]:
+    -   @learncard/vc-plugin@1.2.0
+    -   @learncard/expiration-plugin@1.1.61
+
 ## 2.3.18
 
 ### Patch Changes